(BSR)[API] refactor: move adage serialize in pcapi serialization
