+++ api/src/pcapi/core/educational/adage_backends/__init__.py
@@ -2,17 +2,16 @@ import datetime
 
 from pcapi import settings
 from pcapi.connectors.serialization.api_adage_serializers import AdageVenue
-from pcapi.core.educational import schemas as educational_schemas
-from pcapi.core.educational.adage_backends import serialize
+from pcapi.serialization.educational.adage import shared as adage_serialize
 from pcapi.utils.module_loading import import_string
 
 
-def notify_prebooking(data: educational_schemas.EducationalBookingResponse) -> None:
+def notify_prebooking(data: adage_serialize.EducationalBookingResponse) -> None:
     backend = import_string(settings.ADAGE_BACKEND)
     backend().notify_prebooking(data=data)
 
 
-def notify_offer_or_stock_edition(data: educational_schemas.EducationalBookingEdition) -> None:
+def notify_offer_or_stock_edition(data: adage_serialize.EducationalBookingEdition) -> None:
     backend = import_string(settings.ADAGE_BACKEND)
     backend().notify_offer_or_stock_edition(data=data)
 
@@ -23,7 +22,9 @@ def get_adage_offerer(siren: str) -> list[AdageVenue]:
     return result
 
 
-def notify_booking_cancellation_by_offerer(data: educational_schemas.EducationalBookingResponse) -> None:
+def notify_booking_cancellation_by_offerer(
+    data: adage_serialize.EducationalBookingResponse,
+) -> None:
     backend = import_string(settings.ADAGE_BACKEND)
     backend().notify_booking_cancellation_by_offerer(data=data)
 
@@ -34,18 +35,20 @@ def get_cultural_partners(since_date: datetime.datetime | None = None) -> list[d
     return result
 
 
-def notify_institution_association(data: serialize.AdageCollectiveOffer) -> None:
+def notify_institution_association(data: adage_serialize.AdageCollectiveOffer) -> None:
     backend = import_string(settings.ADAGE_BACKEND)
     backend().notify_institution_association(data=data)
 
 
-def get_cultural_partner(siret: str) -> educational_schemas.AdageCulturalPartner:
+def get_cultural_partner(siret: str) -> adage_serialize.AdageCulturalPartner:
     backend = import_string(settings.ADAGE_BACKEND)
     result = backend().get_cultural_partner(siret)
     return result
 
 
-def get_adage_educational_institutions(ansco: str) -> list[serialize.AdageEducationalInstitution]:
+def get_adage_educational_institutions(
+    ansco: str,
+) -> list[adage_serialize.AdageEducationalInstitution]:
     backend = import_string(settings.ADAGE_BACKEND)
     result = backend().get_adage_educational_institutions(ansco)
     return result
@@ -57,13 +60,17 @@ def get_adage_educational_redactor_from_uai(uai: str) -> list[dict[str, str]]:
     return result
 
 
-def notify_reimburse_collective_booking(data: educational_schemas.AdageReimbursementNotification) -> None:
+def notify_reimburse_collective_booking(
+    data: adage_serialize.AdageReimbursementNotification,
+) -> None:
     backend = import_string(settings.ADAGE_BACKEND)
     result = backend().notify_reimburse_collective_booking(data)
     return result
 
 
-def notify_redactor_when_collective_request_is_made(data: serialize.AdageCollectiveRequest) -> None:
+def notify_redactor_when_collective_request_is_made(
+    data: adage_serialize.AdageCollectiveRequest,
+) -> None:
     backend = import_string(settings.ADAGE_BACKEND)
     result = backend().notify_redactor_when_collective_request_is_made(data)
     return result
