+++ api/src/pcapi/core/educational/adage_backends/testing.py
@@ -1,28 +1,25 @@
 import datetime
 import decimal
 
-import pcapi.core.educational.schemas as educational_schemas
 from pcapi.connectors.serialization.api_adage_serializers import AdageVenue
 from pcapi.core.educational import exceptions
 from pcapi.core.educational.adage_backends.base import AdageClient
-from pcapi.core.educational.adage_backends.serialize import AdageCollectiveOffer
-from pcapi.core.educational.adage_backends.serialize import AdageCollectiveRequest
-from pcapi.core.educational.adage_backends.serialize import AdageEducationalInstitution
+from pcapi.serialization.educational.adage import shared as adage_serialize
 
 from .. import testing
 
 
 class AdageSpyClient(AdageClient):
-    def notify_prebooking(self, data: educational_schemas.EducationalBookingResponse) -> None:
+    def notify_prebooking(self, data: adage_serialize.EducationalBookingResponse) -> None:
         testing.adage_requests.append({"url": f"{self.base_url}/v1/prereservation", "sent_data": data})
 
-    def notify_offer_or_stock_edition(self, data: educational_schemas.EducationalBookingResponse) -> None:
+    def notify_offer_or_stock_edition(self, data: adage_serialize.EducationalBookingResponse) -> None:
         testing.adage_requests.append({"url": f"{self.base_url}/v1/prereservation-edit", "sent_data": data})
 
     def get_adage_offerer(self, siren: str) -> list[AdageVenue]:
         raise RuntimeError("Do not use the spy for this method, mock the get request instead")
 
-    def notify_booking_cancellation_by_offerer(self, data: educational_schemas.EducationalBookingResponse) -> None:
+    def notify_booking_cancellation_by_offerer(self, data: adage_serialize.EducationalBookingResponse) -> None:
         testing.adage_requests.append({"url": f"{self.base_url}/v1/prereservation-annule", "sent_data": data})
 
     def get_cultural_partners(
@@ -93,12 +90,12 @@ class AdageSpyClient(AdageClient):
             },
         ]
 
-    def notify_institution_association(self, data: AdageCollectiveOffer) -> None:
+    def notify_institution_association(self, data: adage_serialize.AdageCollectiveOffer) -> None:
         testing.adage_requests.append({"url": f"{self.base_url}/v1/offre-assoc", "sent_data": data})
 
-    def get_cultural_partner(self, siret: str) -> educational_schemas.AdageCulturalPartner:
+    def get_cultural_partner(self, siret: str) -> adage_serialize.AdageCulturalPartner:
         testing.adage_requests.append({"url": f"{self.base_url}/v1/partenaire-culturel/{siret}", "sent_data": ""})
-        return educational_schemas.AdageCulturalPartner(
+        return adage_serialize.AdageCulturalPartner(
             id=128028,
             venueId=None,
             siret=siret,
@@ -128,10 +125,10 @@ class AdageSpyClient(AdageClient):
             synchroPass=0,
         )
 
-    def get_adage_educational_institutions(self, ansco: str) -> list[AdageEducationalInstitution]:
+    def get_adage_educational_institutions(self, ansco: str) -> list[adage_serialize.AdageEducationalInstitution]:
         testing.adage_requests.append({"url": f"{self.base_url}/v1/etablissement-culturel/{ansco}", "sent_data": ""})
         return [
-            AdageEducationalInstitution(
+            adage_serialize.AdageEducationalInstitution(
                 uai="0470009E",
                 sigle="COLLEGE",
                 libelle="DE LA TOUR0",
@@ -142,7 +139,7 @@ class AdageSpyClient(AdageClient):
                 latitude=decimal.Decimal("48.8534100"),
                 longitude=decimal.Decimal("2.3488000"),
             ),
-            AdageEducationalInstitution(
+            adage_serialize.AdageEducationalInstitution(
                 uai="0470010E",
                 sigle="CLG",
                 libelle="Balamb Garden",
@@ -181,9 +178,9 @@ class AdageSpyClient(AdageClient):
             ]
         raise exceptions.EducationalRedactorNotFound("No educational redactor found for the given UAI")
 
-    def notify_reimburse_collective_booking(self, data: educational_schemas.AdageReimbursementNotification) -> None:
+    def notify_reimburse_collective_booking(self, data: adage_serialize.AdageReimbursementNotification) -> None:
         api_url = f"{self.base_url}/v1/reservation-remboursement"
         testing.adage_requests.append({"url": api_url, "sent_data": ""})
 
-    def notify_redactor_when_collective_request_is_made(self, data: AdageCollectiveRequest) -> None:
+    def notify_redactor_when_collective_request_is_made(self, data: adage_serialize.AdageCollectiveRequest) -> None:
         testing.adage_requests.append({"url": f"{self.base_url}/v1/offre-vitrine", "sent_data": data})
