--- api/src/pcapi/routes/adage/v1/serialization/constants.py
@@ -1,1 +1,0 @@
-EDUCATIONAL_BOOKING_NOT_FOUND = "EDUCATIONAL_BOOKING_NOT_FOUND"
