+++ api/src/pcapi/routes/adage/v1/serialization/educational_institution.py
@@ -1,13 +1,13 @@
 from pydantic.v1.fields import Field
 
-import pcapi.core.educational.schemas as educational_schemas
+from pcapi.serialization.educational.adage import shared as adage_serialize
 from pcapi.serialization.utils import to_camel
 
 
-class EducationalInstitutionResponse(educational_schemas.AdageBaseResponseModel):
+class EducationalInstitutionResponse(adage_serialize.AdageBaseResponseModel):
     credit: int = Field(description="Total credit granted to the educational institution")
     isFinal: bool = Field(description="Flag to know if the credit has been approved and is now final")
-    prebookings: list[educational_schemas.EducationalBookingResponse]
+    prebookings: list[adage_serialize.EducationalBookingResponse]
 
     class Config:
         title = "School response model"
